feat: add estimated APR data to v2 vault snapshots
Adds support for querying and exposing estimated APR data (from CRV, Velo, and Aero protocols) in v2 vault snapshots. Includes new EstimatedApr schema with APR/APY values and component breakdown.
