+++ packages/ingest/db.ts
@@ -2,7 +2,7 @@
 
 import { z } from 'zod'
 import { strings } from 'lib'
-import { EstimatedAprSchema, StrideSchema } from 'lib/types'
+import { StrideSchema } from 'lib/types'
 import { Pool, PoolClient, types as pgTypes } from 'pg'
 import { snakeToCamelCols } from 'lib/strings'
 
@@ -178,62 +178,6 @@ export async function getLatestOracleApr(chainId: number, address: string): Prom
   return [result.apr || 0, result.apy || 0]
 }
 
-export async function getLatestEstimatedApr(chainId: number, address: string) {
-  const result = await firstRow(`
-  SELECT
-    chain_id as "chainId",
-    address,
-    label,
-    MAX(CASE WHEN component = 'netAPR' THEN value END) AS apr,
-    MAX(CASE WHEN component = 'netAPY' THEN value END) AS apy,
-    MAX(CASE WHEN component = 'boost' THEN value END) AS boost,
-    MAX(CASE WHEN component = 'poolAPY' THEN value END) AS "poolAPY",
-    MAX(CASE WHEN component = 'boostedAPR' THEN value END) AS "boostedAPR",
-    MAX(CASE WHEN component = 'baseAPR' THEN value END) AS "baseAPR",
-    MAX(CASE WHEN component = 'rewardsAPR' THEN value END) AS "rewardsAPR",
-    MAX(CASE WHEN component = 'rewardsAPY' THEN value END) AS "rewardsAPY",
-    MAX(CASE WHEN component = 'cvxAPR' THEN value END) AS "cvxAPR",
-    MAX(CASE WHEN component = 'keepCRV' THEN value END) AS "keepCRV",
-    MAX(CASE WHEN component = 'keepVelo' THEN value END) AS "keepVelo",
-    block_number as "blockNumber",
-    block_time as "blockTime"
-  FROM output
-  WHERE block_time = (
-      SELECT MAX(block_time) FROM output
-      WHERE chain_id = $1
-      AND LOWER(address) = LOWER($2)
-      AND label IN ('crv-estimated-apr', 'velo-estimated-apr', 'aero-estimated-apr')
-    )
-    AND chain_id = $1
-    AND LOWER(address) = LOWER($2)
-    AND label IN ('crv-estimated-apr', 'velo-estimated-apr', 'aero-estimated-apr')
-  GROUP BY chain_id, address, label, block_number, block_time;
-  `, [chainId, address])
-
-  if (!result) return undefined
-
-  let type = 'unknown'
-  if (result.label === 'crv-estimated-apr') type = 'crv'
-  if (result.label === 'velo-estimated-apr') type = 'velo'
-  if (result.label === 'aero-estimated-apr') type = 'aero'
-
-  return EstimatedAprSchema.parse({
-    apr: result.apr || 0,
-    apy: result.apy || 0,
-    type,
-    components: {
-      boost: result.boost,
-      poolAPY: result.poolAPY,
-      boostedAPR: result.boostedAPR,
-      baseAPR: result.baseAPR,
-      rewardsAPR: result.rewardsAPR,
-      rewardsAPY: result.rewardsAPY,
-      cvxAPR: result.cvxAPR,
-      keepCRV: result.keepCRV,
-      keepVelo: result.keepVelo
-    }
-  })
-}
 
 export function toUpsertSql(table: string, pk: string, data: object, where?: string) {
   const timestampConversionExceptions = [ 'profit_max_unlock_time' ]
